Change interface of get_profile to hide response details
Previously we returned response.map which is essentially the raw response from
the api. In CAS 5 the shape of this response changed, we don't want clients of
caslib.py to have to react to this. The OAuthResponse class will know how to handle the
different api responses and provide a dependable profile object.